--- src/PiCamModule.py
@@ -1,3 +1,0 @@
-from picamera import PiCamera
-
-camera = PiCamera()
