+++ src/reference/LassoCam.py
