+++ src/reference/camera_test.py
