+++ src/reference/old_lassoCam.py
