+++ src/reference/ptzSmooth.py
